rename `settings` module to `settings-page`
